Update arbitrage.java
This code is checking for arbitrage opportunities in betting on a sports game. It does this by checking the probability of each possible outcome (team 1 winning, team 2 winning, and a draw) at each bookmaker, and seeing if the sum of these probabilities is less than 100%. If it is, then there is an arbitrage opportunity because it is possible to place bets in such a way that, no matter what the outcome of the game is, the return will be greater than the original bet. The code then prints out the bookmakers and the percentage profit opportunity.